+++ CHANGELOG.md
@@ -2,8 +2,37 @@
 
 All notable changes to this project will be documented in this file.
 
-The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
-to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## [3.3.0] - 2024-07-14
+
+### Added
+
+- Type value verification.
+- Load ComicInfo from a `cbz` file.
+- Built-in reader for dynamic viewing.
+- Image compatibility checking.
+- Added `save` method to directly save a page or comic.
+- Display additional file information via the player.
+
+### Changed
+
+- Complete code refactoring.
+- `pages` variable now public for native page manipulation in Python.
+- Optimized `pack` function with caching.
+- Library now supports loading `zip` files containing only images.
+
+### Fixed
+
+- Default value definitions.
+- Corrected values written to XML.
+
+### New Contributors
+
+- [piskunqa](https://github.com/piskunqa)
+- [OleskiiPyskun](https://github.com/OleskiiPyskun)
+- [domenicoblanco](https://github.com/domenicoblanco)
+- [RivMt](https://github.com/RivMt)
 
 ## [3.2.0] - 2024-03-17
 
@@ -53,6 +82,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 - Initial release.
 
+[3.3.0]: https://github.com/hyugogirubato/cbz/releases/tag/v3.3.0
 [3.2.0]: https://github.com/hyugogirubato/cbz/releases/tag/v3.2.0
 [3.1.2]: https://github.com/hyugogirubato/cbz/releases/tag/v3.1.2
 [3.1.1]: https://github.com/hyugogirubato/cbz/releases/tag/v3.1.1